+++ vite.config.js
@@ -10,8 +10,8 @@ export default defineConfig({
   base: './',
   server: {
     proxy: {
-      '/api':     'http://localhost:3003',
-      '/uploads': 'http://localhost:3003',
+      '/api':     'http://localhost:3002',
+      '/uploads': 'http://localhost:3002',
     }
   },
   plugins: [